refactor : Util class => 함수로 변경
