+++ src/js/utils/dom.js
@@ -1,0 +1,7 @@
+export const getQuerySelector = (selector) => {
+  return document.querySelector(selector);
+};
+
+export const setElementDisplay = ($element, displayValue) => {
+  $element.style.display = displayValue;
+};
