+++ src/js/utils/random.js
@@ -1,0 +1,3 @@
+export const generateRandomNumber = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
