--- src/js/utils/util.js
@@ -1,19 +1,0 @@
-class Util {
-  generateRandomNumber(min, max) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-
-  getQuerySelector(selector) {
-    return document.querySelector(selector);
-  }
-
-  setElementDisplay($element, displayValue) {
-    $element.style.display = displayValue;
-  }
-}
-
-export const {
-  generateRandomNumber,
-  getQuerySelector,
-  setElementDisplay,
-} = new Util();
